Moved  Json serialization to Serialization.Json namespace
