+++ source/MonoGame.Extended/Serialization/Json/JsonContentLoader.cs
@@ -3,7 +3,7 @@ using Microsoft.Xna.Framework.Content;
 using MonoGame.Extended.Content;
 
 
-namespace MonoGame.Extended.Serialization
+namespace MonoGame.Extended.Serialization.Json
 {
     public class JsonContentLoader : IContentLoader
     {
