+++ source/MonoGame.Extended/Serialization/Json/ThicknessJsonConverter.cs
@@ -2,7 +2,7 @@ using System;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace MonoGame.Extended.Serialization;
+namespace MonoGame.Extended.Serialization.Json;
 
 /// <summary>
 /// Converts a <see cref="Thickness"/> value to or from JSON.
