+++ source/MonoGame.Extended/Serialization/Json/Utf8JsonReaderExtensions.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Text.Json;
 
-namespace MonoGame.Extended.Serialization;
+namespace MonoGame.Extended.Serialization.Json;
 
 /// <summary>
 /// Provides extension methods for working with <see cref="Utf8JsonReader"/>.
